refactor: break evaluation overview file into sub-components
