+++ app/packages/core/src/plugins/SchemaIO/components/NativeModelEvaluationView/constants.ts
@@ -1,5 +1,3 @@
 export const KEY_COLOR = "#ff6d04";
 export const COMPARE_KEY_COLOR = "#03a9f4";
 export const COMPARE_KEY_SECONDARY_COLOR = "#87D2FA";
-export const DEFAULT_BAR_CONFIG = { sortBy: "default" };
-export const NONE_CLASS = "(none)";
