+++ app/packages/core/src/plugins/SchemaIO/components/NativeModelEvaluationView/evaluation/overview/constants.ts
@@ -1,0 +1,16 @@
+export const CLASS_PERFORMANCE_SORT_OPTIONS = [
+  { value: "default", label: "Default" },
+  { value: "az", label: "Alphabetical (A-Z)" },
+  { value: "za", label: "Alphabetical (Z-A)" },
+  { value: "best", label: "Best performing" },
+  { value: "worst", label: "Worst performing" },
+];
+export const CONFUSION_MATRIX_SORT_OPTIONS = [
+  { value: "default", label: "Default" },
+  { value: "az", label: "Alphabetical (A-Z)" },
+  { value: "za", label: "Alphabetical (Z-A)" },
+  { value: "mc", label: "Most common classes" },
+  { value: "lc", label: "Least common classes" },
+];
+export const DEFAULT_BAR_CONFIG = { sortBy: "default" };
+export const NONE_CLASS = "(none)";
